Consolidate types, catalogue, and classes to one serviceTrait
