+++ server/service/src/catalogue/validate_catalogue.rs
